Warnung weg
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/branches/SGppCodingDays@3999 4eea3252-f0fb-4393-894d-40516dce545b

+++ base/src/sgpp/base/grid/generation/functors/SmoothedErrorRefinementFunctor.cpp
@@ -69,8 +69,8 @@ double SmoothedErrorRefinementFunctor::operator()(GridStorage* storage,
 		return 0.0;
 	}
 
-	for (int i = 0; i < numErrors; i++) {
-		SmoothedErrors += errors[i] * (1.0 / numErrors);
+	for (unsigned int i = 0; i < numErrors; i++) {
+		SmoothedErrors += errors[i] * (1.0 / (double)numErrors);
 	}
 
 	return SmoothedErrors;
